Moved 'layout' files into 'src/components/layout/' dir
