+++ www/src/components/layout/search-form/algolia-styles.js
